ENG-6348 - refactor activity and fragment callbacks into abstract classes in main - move NIDViewsExtension, NIDDataStoreManager, NeuroID out or release and develop variants directories to main
