+++ NeuroID/src/androidLib/java/com/neuroid/tracker/callbacks/NIDActivityCallbacks.kt
@@ -15,18 +15,13 @@ import com.neuroid.tracker.models.NIDEventModel
 import com.neuroid.tracker.service.NIDServiceTracker
 import com.neuroid.tracker.storage.getDataStoreInstance
 import com.neuroid.tracker.utils.NIDLog
+import com.neuroid.tracker.utils.NIDLogWrapper
 import org.json.JSONArray
 import org.json.JSONObject
 
-class NIDActivityCallbacks: ActivityLifecycleCallbacks {
+class NIDActivityCallbacks: ActivityCallbacks() {
     private var auxOrientation = -1
-    private var activitiesStarted = 0
-    private var listActivities = ArrayList<String>()
     private var wasChanged = false
-
-    override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-        NIDLog.d("Neuro ID", "NIDDebug onActivityCreated");
-    }
 
     override fun onActivityStarted(activity: Activity) {
         NIDLog.d("NID--Activity", "Activity - Created")
@@ -71,7 +66,6 @@ class NIDActivityCallbacks: ActivityLifecycleCallbacks {
             auxOrientation = orientation
         }
 
-
         val metadataObj = JSONObject()
         metadataObj.put("component", "activity")
         metadataObj.put("lifecycle", "postCreated")
@@ -90,22 +84,7 @@ class NIDActivityCallbacks: ActivityLifecycleCallbacks {
             )
     }
 
-    /**
-     * Option for customers to force start with Activity
-     */
-    fun forceStart(activity: Activity) {
-        registerTargetFromScreen(
-            activity,
-            activityOrFragment = "activity",
-            parent = activity::class.java.simpleName
-        )
-        // register listeners for focus, blur and touch events
-        registerWindowListeners(activity)
-    }
-
     override fun onActivityResumed(activity: Activity) {
-        //No op
-        // SHOULD BE WINDOW FOCUS?
         NIDLog.d("NID--Activity", "Activity - Resumed")
 
         val gyroData = NIDSensorHelper.getGyroscopeInfo()
@@ -116,6 +95,7 @@ class NIDActivityCallbacks: ActivityLifecycleCallbacks {
         metadataObj.put("lifecycle", "resumed")
         metadataObj.put("className", "${activity::class.java.simpleName}")
         val attrJSON = JSONArray().put(metadataObj)
+
         getDataStoreInstance()
             .saveEvent(
                 NIDEventModel(
@@ -133,6 +113,8 @@ class NIDActivityCallbacks: ActivityLifecycleCallbacks {
 
         registerTargetFromScreen(
             activity,
+            NIDLogWrapper(),
+            getDataStoreInstance(),
             registerTarget = true,
             registerListeners = true,
             activityOrFragment = "activity",
@@ -140,68 +122,6 @@ class NIDActivityCallbacks: ActivityLifecycleCallbacks {
         )
 
         registerWindowListeners(activity)
-
     }
 
-    override fun onActivityPaused(activity: Activity) {
-        //No op
-        // SHOULD BE WINDOW BLUR?
-        NIDLog.d("NID--Activity", "Activity - Paused")
-
-        val gyroData = NIDSensorHelper.getGyroscopeInfo()
-        val accelData = NIDSensorHelper.getAccelerometerInfo()
-
-        val metadataObj = JSONObject()
-        metadataObj.put("component", "activity")
-        metadataObj.put("lifecycle", "paused")
-        metadataObj.put("className", "${activity::class.java.simpleName}")
-        val attrJSON = JSONArray().put(metadataObj)
-
-        getDataStoreInstance()
-            .saveEvent(
-                NIDEventModel(
-                    type = WINDOW_BLUR,
-                    ts = System.currentTimeMillis(),
-                    gyro = gyroData,
-                    accel = accelData,
-                    attrs = attrJSON
-                )
-            )
-    }
-
-    override fun onActivityStopped(activity: Activity) {
-        NIDLog.d("NID--Activity", "Activity - Stopped")
-        activitiesStarted--
-    }
-
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
-        // No Operation
-        NIDLog.d("NID--Activity", "Activity - Save Instance")
-    }
-
-    override fun onActivityDestroyed(activity: Activity) {
-        NIDLog.d("NID--Activity", "Activity - Destroyed")
-        val gyroData = NIDSensorHelper.getGyroscopeInfo()
-        val accelData = NIDSensorHelper.getAccelerometerInfo()
-        val activityDestroyed = activity::class.java.name
-        listActivities.remove(activityDestroyed)
-
-        val metadataObj = JSONObject()
-        metadataObj.put("component", "activity")
-        metadataObj.put("lifecycle", "destroyed")
-        metadataObj.put("className", "${activity::class.java.simpleName}")
-        val attrJSON = JSONArray().put(metadataObj)
-
-        NIDLog.d("NID--Activity", "Activity - Destroyed - Window Unload")
-        getDataStoreInstance()
-            .saveEvent(
-                NIDEventModel(
-                    type = WINDOW_UNLOAD,
-                    ts = System.currentTimeMillis(),
-                    gyro = gyroData,
-                    accel = accelData,
-                    attrs = attrJSON
-                )
-            )
-    }
 }
