+++ NeuroID/src/main/java/com/neuroid/tracker/utils/NIDTextWatcher.kt
@@ -1,6 +1,5 @@
 package com.neuroid.tracker.utils
 
-import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
 import android.text.Editable
