+++ app/src/androidTest/java/com/sample/neuroid/us/NIDSchema.kt
@@ -15,7 +15,6 @@ import org.json.JSONArray
 import org.json.JSONObject
 import org.junit.Assert
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import java.io.BufferedReader
 import java.io.IOException
